ADD: JwtConfig class, JwtAuthFilter class, TokenService
Made a JwtConfig class, to fix circular dependency. So i moved the jwt stuff in SecurityConfig to JwtConfig.

JwtAuthFilter - A JWT token filter.

MyUserDetailsService - method that adds the roles was wrong and was fixed.

TokenService - Some JWT methods to get username and validate jwt token

Page.tsx in register - small fix in text.

Tested the jwtfilter with postman and it worked.