ADD: navbar changes + module.css
Made the navbar changes in layout to be in the main homepage, or else it would also be in the login and register pages.

made a .module.css for login will make one for register and logout, it fixes a bug that if you login and redirects to homepage the css will transfer over too, with .module it does not happen.