ADD: Testing with WebSocket
Added WebSocket dependencies to backend and frontend.

Went with double auth for WebSocket, one before the connection and the other after the connection(every message*), so if the jwt is invalid you will not be able to send messages.

Made a test page that tests the new WebSocket controller, it sent the message to the db.