ADD: WebSocket receive message
Made it so we use the message object the backend sends and based on the selecteduser you either see the new message or get a notification.

Also removed sending the message to yourself from the backend, it was easier to implement on the frontend, so you dont have to go back and forth.